refactor(core): simplify rendering pipeline and build configuration
- Remove `layeredCanvasRenderer.ts` and `tdollRenderCache.service.ts`
- Extract `replacedQueryMatch` into dedicated `query.ts` module
- Introduce `tsconfig.build.json` for build-specific exclusions
- Externalize runtime dependencies and Node builtins in Rollup config
