--- src/commands/tdoll/utils/cachedRenderer.ts
@@ -1,237 +1,0 @@
-/**
- * CachedTDollRenderer - TDoll 分层渲染缓存实现
- *
- * 功能:
- * 1. 主内容缓存 - 缓存渲染的主要图像内容
- * 2. 动态时间戳 - 每次请求时重新叠加时间戳
- * 3. 智能缓存管理 - TTL + LRU 策略
- */
-
-import {
-    createCanvas,
-    Canvas2DContext,
-    CanvasLike,
-} from '../../../services/canvasBackend';
-import { logger } from '../../../utils/logger';
-import {
-    TDollCacheParams,
-    getTDollRenderCache,
-} from '../../../services/tdollRenderCache.service';
-import { OUTPUT_FOLDER } from '../types/constants';
-import * as fs from 'node:fs';
-import * as path from 'node:path';
-
-export interface LayeredRenderResult {
-    /** 最终输出文件路径 */
-    outputPath: string;
-    /** 是否使用了缓存 */
-    fromCache: boolean;
-    /** 缓存命中次数（仅当使用缓存时） */
-    cacheHitCount?: number;
-}
-
-/**
- * 分层渲染基类
- * 提供缓存管理和分层渲染的基础设施
- */
-export abstract class CachedTDollRenderer {
-    protected query: string;
-    protected fileName: string;
-    protected cacheParams: TDollCacheParams;
-    protected dimensions: { width: number; height: number };
-
-    constructor(
-        query: string,
-        fileName: string,
-        cacheParams: TDollCacheParams,
-        dimensions: { width: number; height: number },
-    ) {
-        this.query = query;
-        this.fileName = fileName;
-        this.cacheParams = cacheParams;
-        this.dimensions = dimensions;
-    }
-
-    /**
-     * 渲染主内容（子类实现）
-     * 这部分会被缓存
-     */
-    protected abstract renderContent(
-        context: Canvas2DContext,
-        width: number,
-        height: number,
-    ): Promise<void>;
-
-    /**
-     * 渲染时间戳（子类实现）
-     * 这部分每次都会重新渲染
-     */
-    protected abstract renderTimestamp(
-        context: Canvas2DContext,
-        width: number,
-        height: number,
-    ): Promise<void>;
-
-    /**
-     * 获取时间戳区域高度
-     */
-    protected abstract getTimestampHeight(): number;
-
-    /**
-     * 执行渲染（带缓存）
-     */
-    async render(): Promise<LayeredRenderResult> {
-        const cache = getTDollRenderCache();
-        const outputDir = path.join(process.cwd(), OUTPUT_FOLDER);
-        const outputPath = path.join(outputDir, this.fileName);
-
-        // 确保输出目录存在
-        if (!fs.existsSync(outputDir)) {
-            fs.mkdirSync(outputDir, { recursive: true });
-        }
-
-        // 检查缓存
-        if (cache) {
-            const cachedEntry = cache.get(this.cacheParams);
-            if (cachedEntry) {
-                // 缓存命中，使用缓存的主内容
-                logger.info(
-                    `[CachedTDollRenderer] Cache hit for ${this.query}, reusing content`,
-                );
-
-                // 读取缓存的内容图片
-                const contentPath = cachedEntry.contentPath;
-                if (fs.existsSync(contentPath)) {
-                    // 创建新画布，叠加时间戳
-                    const finalCanvas =
-                        await this.composeWithTimestamp(contentPath);
-
-                    // 保存最终图片
-                    this.saveCanvas(finalCanvas, outputPath);
-
-                    return {
-                        outputPath,
-                        fromCache: true,
-                        cacheHitCount: cachedEntry.accessCount,
-                    };
-                }
-            }
-        }
-
-        // 缓存未命中，执行完整渲染
-        logger.info(
-            `[CachedTDollRenderer] Cache miss for ${this.query}, rendering fresh`,
-        );
-
-        // 1. 渲染主内容
-        const contentCanvas = createCanvas(
-            this.dimensions.width,
-            this.dimensions.height - this.getTimestampHeight(),
-        );
-        const contentContext = contentCanvas.getContext('2d');
-        await this.renderContent(
-            contentContext,
-            this.dimensions.width,
-            this.dimensions.height - this.getTimestampHeight(),
-        );
-
-        // 2. 保存主内容（用于缓存）
-        const contentFileName = `content_${this.fileName}`;
-        const contentPath = path.join(outputDir, contentFileName);
-        this.saveCanvas(contentCanvas, contentPath);
-
-        // 3. 叠加时间戳并保存最终图片
-        const finalCanvas = createCanvas(
-            this.dimensions.width,
-            this.dimensions.height,
-        );
-        const finalContext = finalCanvas.getContext('2d');
-
-        // 绘制主内容
-        finalContext.drawImage(contentCanvas as any, 0, 0);
-
-        // 绘制时间戳
-        await this.renderTimestamp(
-            finalContext,
-            this.dimensions.width,
-            this.getTimestampHeight(),
-        );
-
-        this.saveCanvas(finalCanvas, outputPath);
-
-        // 4. 更新缓存
-        if (cache) {
-            cache.set(
-                this.cacheParams,
-                contentPath,
-                outputPath,
-                this.dimensions,
-            );
-        }
-
-        return {
-            outputPath,
-            fromCache: false,
-        };
-    }
-
-    /**
-     * 将缓存的主内容与新的时间戳合成
-     */
-    private async composeWithTimestamp(
-        contentPath: string,
-    ): Promise<CanvasLike> {
-        // 注意：这里需要加载图片
-        // 由于 canvas 限制，我们重新渲染主内容
-        // 实际实现中，可以优化为直接加载图片文件
-
-        const finalCanvas = createCanvas(
-            this.dimensions.width,
-            this.dimensions.height,
-        );
-        const finalContext = finalCanvas.getContext('2d');
-
-        // 重新渲染主内容
-        const contentHeight =
-            this.dimensions.height - this.getTimestampHeight();
-        await this.renderContent(
-            finalContext,
-            this.dimensions.width,
-            contentHeight,
-        );
-
-        // 渲染时间戳
-        await this.renderTimestamp(
-            finalContext,
-            this.dimensions.width,
-            this.getTimestampHeight(),
-        );
-
-        return finalCanvas;
-    }
-
-    /**
-     * 保存画布到文件
-     */
-    private saveCanvas(canvas: CanvasLike, filePath: string): void {
-        const buffer = canvas.toBuffer('image/png');
-        fs.writeFileSync(filePath, buffer);
-    }
-}
-
-/**
- * 便捷函数：创建缓存参数
- */
-export function createCacheParams(
-    query: string,
-    commandType: 'tdoll' | 'tdollskin',
-    query2?: string,
-    tdollId?: string,
-): TDollCacheParams {
-    return {
-        query,
-        query2,
-        tdollId,
-        commandType,
-    };
-}
