+++ src/commands/tdoll/utils/utils.ts
@@ -55,13 +55,7 @@ export const getRandomTDollData = (dataList: ITDollDataItem[]): string => {
     return formatTDollData(randomData);
 };
 
-export const replacedQueryMatch = (query: string): string => {
-    return query
-        .toLowerCase()
-        .replaceAll('-', '')
-        .replaceAll('.', '')
-        .replaceAll(' ', '');
-};
+export { replacedQueryMatch } from './query';
 
 /**
  * Get matched tdoll data
